requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
deep-translator==1.11.4
frozenlist==1.6.0
idna==3.10
lxml==5.4.0
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
